docs(example): example app with new controller model
