--- example/lib/widgets/widgets.dart
@@ -1,2 +1,0 @@
-export 'simple_decorated_box.dart';
-export 'rich_decorated_box.dart';
